requirements.txt
and==66.0.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
numpy==1.24.2
pandas==1.5.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
zipp==3.15.0

django-debug-toolbar~=3.8.1